requirements.txt
tensorflow
matplotlib
numpy
pillow
scikit-learn
scipy   